+++ glade_forms/lib/src/core/error/error_serverity.dart
@@ -1,4 +1,0 @@
-enum ErrorServerity {
-  error,
-  warning,
-}