+++ glade_forms/lib/src/core/input/glade_string_input.dart
@@ -1,4 +1,4 @@
-import 'package:glade_forms/src/core/error/glade_error_keys.dart';
+import 'package:glade_forms/src/core/error/glade_validations_keys.dart';
 import 'package:glade_forms/src/core/input/glade_input.dart';
 import 'package:glade_forms/src/validator/part/satisfy_predicate_part.dart';
 import 'package:glade_forms/src/validator/specialized/string_validator.dart';
@@ -39,7 +39,7 @@ class GladeStringInput extends GladeInput<String> {
 
   /// Maximum length of the string if `maxLength()` standard validator is used.
   int? getMaxLength({Object? key}) {
-    final validator = validatorInstance.tryFindValidatorPart(key ?? GladeErrorKeys.stringMaxLength);
+    final validator = validatorInstance.tryFindValidatorPart(key ?? GladeValidationsKeys.stringMaxLength);
 
     if (validator == null) return null;