+++ glade_forms/lib/src/validator/part/input_validator_part.dart
@@ -1,5 +1,5 @@
 import 'package:equatable/equatable.dart';
-import 'package:glade_forms/src/core/error/error_serverity.dart';
+import 'package:glade_forms/src/core/error/validation_severity.dart';
 import 'package:glade_forms/src/validator/validator_result/glade_validator_result.dart';
 
 typedef ShouldValidateCallback<T> = bool Function(T value);
@@ -11,13 +11,13 @@ abstract class InputValidatorPart<T> extends Equatable {
   // ignore: prefer-correct-callback-field-name, name is ok.
   final ShouldValidateCallback<T>? shouldValidate;
 
-  final ErrorServerity serverity;
+  final ValidationSeverity serverity;
 
   @override
   // ignore: list-all-equatable-fields, on purpose
   List<Object?> get props => [key, serverity];
 
-  const InputValidatorPart({this.key, this.shouldValidate, this.serverity = ErrorServerity.error});
+  const InputValidatorPart({this.key, this.shouldValidate, this.serverity = ValidationSeverity.error});
 
   GladeValidatorResult<T>? validate(T value);
 }