+++ glade_forms/lib/src/validator/specialized/int_validator.dart
@@ -11,21 +11,21 @@ class IntValidator extends GladeValidator<int> {
   ///
   /// With [inclusiveInterval] set to true(default), the comparison is inclusive.
   ///
-  /// Default [key] is [ GladeErrorKeys.intCompareError].
+  /// Default [key] is [ GladeValidationsKeys.intCompareError].
   void isBetween({
     required int min,
     required int max,
-    OnValidate<int>? devError,
+    OnValidate<int>? devMessage,
     Object? key,
     ShouldValidateCallback<int>? shouldValidate,
     bool inclusiveInterval = true,
-    ErrorServerity severity = ErrorServerity.error,
+    ValidationSeverity severity = ValidationSeverity.error,
   }) =>
       satisfy(
         (value) => inclusiveInterval ? value >= min && value <= max : value > min && value < max,
-        devError: devError ??
+        devMessage: devMessage ??
             (value) => 'Value $value  (inclusiveInterval: $inclusiveInterval) is not in between $min and $max.',
-        key: key ?? GladeErrorKeys.intCompareError,
+        key: key ?? GladeValidationsKeys.intCompareError,
         shouldValidate: shouldValidate,
         severity: severity,
       );
@@ -34,19 +34,19 @@ class IntValidator extends GladeValidator<int> {
   ///
   /// If [isInclusive] is set to true(default), the comparison is inclusive.
   ///
-  /// Default [key] is [ GladeErrorKeys.intCompareMinError].
+  /// Default [key] is [ GladeValidationsKeys.intCompareMinError].
   void isMin({
     required int min,
     bool isInclusive = true,
-    OnValidate<int>? devError,
+    OnValidate<int>? devMessage,
     Object? key,
     ShouldValidateCallback<int>? shouldValidate,
-    ErrorServerity severity = ErrorServerity.error,
+    ValidationSeverity severity = ValidationSeverity.error,
   }) =>
       satisfy(
         (value) => isInclusive ? value >= min : value > min,
-        devError: devError ?? (value) => 'Value $value is less than $min.',
-        key: key ?? GladeErrorKeys.intCompareMinError,
+        devMessage: devMessage ?? (value) => 'Value $value is less than $min.',
+        key: key ?? GladeValidationsKeys.intCompareMinError,
         shouldValidate: shouldValidate,
         severity: severity,
       );
@@ -55,19 +55,19 @@ class IntValidator extends GladeValidator<int> {
   ///
   /// If [isInclusive] is set to true(default), the comparison is inclusive.
   ///
-  /// Default [key] is [ GladeErrorKeys.intCompareMaxError].
+  /// Default [key] is [ GladeValidationsKeys.intCompareMaxError].
   void isMax({
     required int max,
     bool isInclusive = true,
-    OnValidate<int>? devError,
+    OnValidate<int>? devMessage,
     Object? key,
     ShouldValidateCallback<int>? shouldValidate,
-    ErrorServerity severity = ErrorServerity.error,
+    ValidationSeverity severity = ValidationSeverity.error,
   }) =>
       satisfy(
         (value) => isInclusive ? value <= max : value < max,
-        devError: devError ?? (value) => 'Value $value is bigger than $max.',
-        key: key ?? GladeErrorKeys.intCompareMaxError,
+        devMessage: devMessage ?? (value) => 'Value $value is bigger than $max.',
+        key: key ?? GladeValidationsKeys.intCompareMaxError,
         shouldValidate: shouldValidate,
         severity: severity,
       );
@@ -76,19 +76,19 @@ class IntValidator extends GladeValidator<int> {
   ///
   /// If [includeZero] is set to true(default), the value can be zero.
   ///
-  /// Default [key] is [GladeErrorKeys.intCompareIsPositiveError].
+  /// Default [key] is [GladeValidationsKeys.intCompareIsPositiveError].
   void isPositive({
     bool includeZero = true,
-    OnValidate<int>? devError,
+    OnValidate<int>? devMessage,
     Object? key,
     ShouldValidateCallback<int>? shouldValidate,
-    ErrorServerity severity = ErrorServerity.error,
+    ValidationSeverity severity = ValidationSeverity.error,
   }) =>
       isMin(
         min: 0,
         isInclusive: includeZero,
-        devError: devError,
-        key: key ?? GladeErrorKeys.intCompareIsPositiveError,
+        devMessage: devMessage,
+        key: key ?? GladeValidationsKeys.intCompareIsPositiveError,
         shouldValidate: shouldValidate,
         severity: severity,
       );
@@ -97,19 +97,19 @@ class IntValidator extends GladeValidator<int> {
   ///
   /// If [includeZero] is set to true(default), the value can be zero.
   ///
-  /// Default [key] is [GladeErrorKeys.intCompareIsNegativeError].
+  /// Default [key] is [GladeValidationsKeys.intCompareIsNegativeError].
   void isNegative({
     bool includeZero = true,
-    OnValidate<int>? devError,
+    OnValidate<int>? devMessage,
     Object? key,
     ShouldValidateCallback<int>? shouldValidate,
-    ErrorServerity severity = ErrorServerity.error,
+    ValidationSeverity severity = ValidationSeverity.error,
   }) =>
       isMax(
         max: 0,
         isInclusive: includeZero,
-        devError: devError,
-        key: key ?? GladeErrorKeys.intCompareIsNegativeError,
+        devMessage: devMessage,
+        key: key ?? GladeValidationsKeys.intCompareIsNegativeError,
         shouldValidate: shouldValidate,
         severity: severity,
       );
@@ -121,15 +121,15 @@ class IntValidatorNullable extends GladeValidator<int?> {
   ///
   /// With [inclusiveInterval] set to true(default), the comparison is inclusive.
   ///
-  /// Default [key] is [ GladeErrorKeys.intCompareError].
+  /// Default [key] is [ GladeValidationsKeys.intCompareError].
   void isBetween({
     required int min,
     required int max,
-    OnValidate<int?>? devError,
+    OnValidate<int?>? devMessage,
     Object? key,
     ShouldValidateCallback<int?>? shouldValidate,
     bool inclusiveInterval = true,
-    ErrorServerity severity = ErrorServerity.error,
+    ValidationSeverity severity = ValidationSeverity.error,
   }) =>
       satisfy(
         (value) {
@@ -137,10 +137,10 @@ class IntValidatorNullable extends GladeValidator<int?> {
 
           return inclusiveInterval ? value >= min && value <= max : value > min && value < max;
         },
-        devError: devError ??
+        devMessage: devMessage ??
             (value) =>
                 'Value ${value ?? 'NULL'} (inclusiveInterval: $inclusiveInterval) is not in between $min and $max.',
-        key: key ?? GladeErrorKeys.intCompareError,
+        key: key ?? GladeValidationsKeys.intCompareError,
         shouldValidate: shouldValidate,
         severity: severity,
       );
@@ -149,14 +149,14 @@ class IntValidatorNullable extends GladeValidator<int?> {
   ///
   /// If [isInclusive] is set to true(default), the comparison is inclusive.
   ///
-  /// Default [key] is [ GladeErrorKeys.intCompareMinError].
+  /// Default [key] is [ GladeValidationsKeys.intCompareMinError].
   void isMin({
     required int min,
     bool isInclusive = true,
-    OnValidate<int?>? devError,
+    OnValidate<int?>? devMessage,
     Object? key,
     ShouldValidateCallback<int?>? shouldValidate,
-    ErrorServerity severity = ErrorServerity.error,
+    ValidationSeverity severity = ValidationSeverity.error,
   }) =>
       satisfy(
         (value) {
@@ -164,8 +164,8 @@ class IntValidatorNullable extends GladeValidator<int?> {
 
           return isInclusive ? value >= min : value > min;
         },
-        devError: devError ?? (value) => 'Value ${value ?? 'NULL'} is less than $min.',
-        key: key ?? GladeErrorKeys.intCompareMinError,
+        devMessage: devMessage ?? (value) => 'Value ${value ?? 'NULL'} is less than $min.',
+        key: key ?? GladeValidationsKeys.intCompareMinError,
         shouldValidate: shouldValidate,
         severity: severity,
       );
@@ -174,14 +174,14 @@ class IntValidatorNullable extends GladeValidator<int?> {
   ///
   /// If [isInclusive] is set to true(default), the comparison is inclusive.
   ///
-  /// Default [key] is [ GladeErrorKeys.intCompareMaxError].
+  /// Default [key] is [ GladeValidationsKeys.intCompareMaxError].
   void isMax({
     required int max,
     bool isInclusive = true,
-    OnValidate<int?>? devError,
+    OnValidate<int?>? devMessage,
     Object? key,
     ShouldValidateCallback<int?>? shouldValidate,
-    ErrorServerity severity = ErrorServerity.error,
+    ValidationSeverity severity = ValidationSeverity.error,
   }) =>
       satisfy(
         (value) {
@@ -189,8 +189,8 @@ class IntValidatorNullable extends GladeValidator<int?> {
 
           return isInclusive ? value <= max : value < max;
         },
-        devError: devError ?? (value) => 'Value ${value ?? 'NULL'} is bigger than $max.',
-        key: key ?? GladeErrorKeys.intCompareMaxError,
+        devMessage: devMessage ?? (value) => 'Value ${value ?? 'NULL'} is bigger than $max.',
+        key: key ?? GladeValidationsKeys.intCompareMaxError,
         shouldValidate: shouldValidate,
         severity: severity,
       );
@@ -199,19 +199,19 @@ class IntValidatorNullable extends GladeValidator<int?> {
   ///
   /// If [includeZero] is set to true(default), the value can be zero.
   ///
-  /// Default [key] is [GladeErrorKeys.intCompareIsPositiveError].
+  /// Default [key] is [GladeValidationsKeys.intCompareIsPositiveError].
   void isPositive({
     bool includeZero = true,
-    OnValidate<int?>? devError,
+    OnValidate<int?>? devMessage,
     Object? key,
     ShouldValidateCallback<int?>? shouldValidate,
-    ErrorServerity severity = ErrorServerity.error,
+    ValidationSeverity severity = ValidationSeverity.error,
   }) =>
       isMin(
         min: 0,
         isInclusive: includeZero,
-        devError: devError,
-        key: key ?? GladeErrorKeys.intCompareIsPositiveError,
+        devMessage: devMessage,
+        key: key ?? GladeValidationsKeys.intCompareIsPositiveError,
         shouldValidate: shouldValidate,
         severity: severity,
       );
@@ -220,19 +220,19 @@ class IntValidatorNullable extends GladeValidator<int?> {
   ///
   /// If [includeZero] is set to true(default), the value can be zero.
   ///
-  /// Default [key] is [GladeErrorKeys.intCompareIsNegativeError].
+  /// Default [key] is [GladeValidationsKeys.intCompareIsNegativeError].
   void isNegative({
     bool includeZero = true,
-    OnValidate<int?>? devError,
+    OnValidate<int?>? devMessage,
     Object? key,
     ShouldValidateCallback<int?>? shouldValidate,
-    ErrorServerity severity = ErrorServerity.error,
+    ValidationSeverity severity = ValidationSeverity.error,
   }) =>
       isMax(
         max: 0,
         isInclusive: includeZero,
-        devError: devError,
-        key: key ?? GladeErrorKeys.intCompareIsNegativeError,
+        devMessage: devMessage,
+        key: key ?? GladeValidationsKeys.intCompareIsNegativeError,
         shouldValidate: shouldValidate,
         severity: severity,
       );